+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## Technical improvement
 
+- DAPI-19: Update `Akeneo\Pim\Enrichment\Component\Product\Normalizer\Indexing\ProductAndProductModel\ProductPropertiesNormalizer` to accept optional normalizers
+
 ## BC breaks
 
 - Twig extension `get_attribute_label_from_code` has been removed with the class `Akeneo\Pim\Structure\Bundle\Twig\AttributeExtension`.